change complete-pattern.js to complete-pattern.ts
