+++ src/test/js/jan/hand-util-test.js
@@ -16,7 +16,7 @@ limitations under the License.
 import assert from 'assert';
 
 import {CompleteInfo} from '../../../main/js/jan/complete-info';
-import CompletePattern from '../../../main/js/jan/complete-pattern';
+import {CompletePattern} from '../../../main/js/jan/complete-pattern';
 import Hand from '../../../main/js/jan/hand';
 import HandUtil from '../../../main/js/jan/hand-util';
 import {Janpai} from '../../../main/js/jan/janpai';
