+++ src/test/js/jan/zjm-util-test.js
@@ -15,7 +15,7 @@ limitations under the License.
  */
 import assert from 'assert';
 
-import CompletePattern from '../../../main/js/jan/complete-pattern';
+import {CompletePattern} from '../../../main/js/jan/complete-pattern';
 import {Janpai} from '../../../main/js/jan/janpai';
 import {JANPAI_ID} from '../../../main/js/jan/janpai-id';
 import {Mentsu} from '../../../main/js/jan/mentsu';
